requirements.txt
web3>=7.0.0
streamlit>=1.40.0
pandas>=2.2.3
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
